Bugfix on stage position importing
Only affected the sliders tab (which may be deprecated soon anyway); the
spinner number model had changed to Doubles instead of Integers, but the
cast was still to Integer.

+++ src/main/java/spim/SPIMAcquisition.java
@@ -1101,7 +1101,7 @@ public class SPIMAcquisition implements MMPlugin, MouseMotionListener, KeyListen
 	protected ActionListener importStagePosition = new ActionListener() {
 		@Override
 		public void actionPerformed(ActionEvent ae) {
-			double range = ((Integer)acqFetchDelta.getValue()).doubleValue();
+			double range = ((Double)acqFetchDelta.getValue()).doubleValue();
 			double value = 0;
 
 			RangeSlider target = null;
